backend/todolist/requirements.txt
Django>=4.2
djangorestframework
celery
redis
django-celery-beat
psycopg2-binary
requests==2.31.0
python-dotenv==1.0.1